Extract downloader into its own file
